+++ components/layout/checkAccount/index.ts
@@ -1,0 +1,1 @@
+export * from "../../check-account/check-account-form";
